+++ TravelPlus_backend/travelPlus/src/main/java/com/travelPlus/v1/Repo/OfferRepo.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OfferRepo extends JpaRepository <Offer,Integer>{
+public interface OfferRepo extends JpaRepository <Offer,Long>{
     List<Offer> findByContract_contractId(int contractId);
 }
